requirements.txt
# Framework
Django==1.11.10
djangorestframework==3.7.7
django-filter==1.1.0
djangorestframework-bulk==0.2.1

# 文档
django-rest-swagger==2.1.2
coreapi==2.3.3
Markdown==2.6.11

# Database
psycopg2==2.7.3.2
redis==2.10.6

# 运维
ansible==2.5.1
paramiko==2.4.1

# Utils
django-cors-headers==2.1.0
Pillow==5.0.0
